project/requirements.txt
wheel
Flask
flask-login
flask-sqlalchemy
flask-migrate
pyopenssl
psycopg2-binary
azure-identity
azure-storage-blob
python-dotenv